RegularQuestionManager, Singleton DAO

+++ app/src/main/java/com/example/dpene/database/model/RegularQuestion.java
@@ -3,11 +3,11 @@ package com.example.dpene.database.model;
 public class RegularQuestion extends Question {
 
     private String[] wrongAnswers;
-    private long levelId;
-    private long nextQuestion;
-    private long questionId; //it could be needed, if not we'll remove it
+    private int levelId;
+    private Integer nextQuestion;
+    private int questionId; //it could be needed, if not we'll remove it
 
-    public RegularQuestion(String question, String rightAnswer, String wrong1, String wrong2, String wrong3, long levelId, long nextQuestion){
+    public RegularQuestion(String question, String rightAnswer, String wrong1, String wrong2, String wrong3, int levelId, Integer nextQuestion){
         super(question, rightAnswer);
         this.wrongAnswers = new String[3];
         this.wrongAnswers[0] = wrong1;
@@ -21,15 +21,15 @@ public class RegularQuestion extends Question {
         return this.wrongAnswers;
     }
 
-    public long getLevelId(){
+    public int getLevelId(){
         return this.levelId;
     }
 
-    public long getNextQuestion(){
+    public Integer getNextQuestion(){
         return nextQuestion;
     }
 
-    public void setQuestionId(long id){
+    public void setQuestionId(int id){
         this.questionId = id;
     }
 }
